refactor: organize frontend and move shared code
