+++ papers2code-ui/src/App.tsx
@@ -3,12 +3,12 @@ import React, { useState, useEffect, useRef } from 'react';
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
 import PaperListPage from './pages/PaperListPage';
 import PaperDetailPage from './pages/PaperDetailPage';
-import logo from './images/papers2codelogo.png';
-import { checkCurrentUser, redirectToGitHubLogin, logoutUser, fetchAndStoreCsrfToken } from './services/auth';
-import type { UserProfile } from './types/user';
-import { UserAvatar } from './components/common';
-import { ModalProvider } from './context/ModalContext'; // Import ModalProvider
-import LoginPromptModal from './components/common/LoginPromptModal'; // Import LoginPromptModal
+import logo from './assets/images/papers2codelogo.png';
+import { checkCurrentUser, redirectToGitHubLogin, logoutUser, fetchAndStoreCsrfToken } from './common/services/auth';
+import type { UserProfile } from './common/types/user';
+import { UserAvatar } from './common/components';
+import { ModalProvider } from './common/context/ModalContext'; // Import ModalProvider
+import LoginPromptModal from './common/components/LoginPromptModal'; // Import LoginPromptModal
 import './App.css';
 
 function App() {
