+++ papers2code-ui/src/common/components/UserAvatar.tsx
