+++ papers2code-ui/src/components/paperDetail/Tabs/ImplementationVoting/ImplementabilityNotice.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { FaThumbsUp, FaThumbsDown } from 'react-icons/fa';
-import { Paper } from '../../../../types/paper';
+import { Paper } from '../../../../common/types/paper';
 import './ImplementabilityNotice.css';
 
 interface ImplementabilityNoticeProps {
