+++ papers2code-ui/src/components/paperDetail/Tabs/ImplementationVoting/ImplementabilityVotingTab.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { Paper, ImplementabilityAction } from '../../../../types/paper';
-import type { UserProfile } from '../../../../types/user';
+import { Paper, ImplementabilityAction } from '../../../../common/types/paper';
+import type { UserProfile } from '../../../../common/types/user';
 import { VoteButton, RetractVoteButton, FaThumbsUp, FaThumbsDown } from '../../VotingButtons';
 import { UserDisplayList } from '../../UserDisplayList'; // Named import
 import './ImplementabilityVotingTab.css';
-import type { PaperActionUsers } from '../../../../services/api'; // Add this line
+import type { PaperActionUsers } from '../../../../common/services/api'; // Add this line
 
 
 interface ImplementabilityTabProps {
