+++ papers2code-ui/src/components/paperDetail/Tabs/Paper/PaperMetadata.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Paper } from '../../../../types/paper'; // Import Status
+import { Paper } from '../../../../common/types/paper'; // Import Status
 import './PaperMetadata.css';
-import { getStatusClass, getStatusSymbol } from '../../../../utils/statusUtils'; // Import utility function for status symbol
+import { getStatusClass, getStatusSymbol } from '../../../../common/utils/statusUtils';
 interface PaperMetadataProps {
     paper: Paper;
 }
