+++ papers2code-ui/src/components/paperList/Search/AdvancedSearchForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
-import { AdvancedPaperFilters } from '../../../services/api';
+import { AdvancedPaperFilters } from '../../../common/services/api';
 import './AdvancedSearchForm.css';
